Move .env to resources directory.
The reason for this is so that we can skip including the file explicitly
in the pubspec.yaml, so that developers don't have to create the file to
get the app working locally. I do wish there was a way to have an
optional, root level asset. Alas, doesn't seem possible.