Depend on specific git version of rustfk.
We're at a point now where builds can easily fail by pulling down the
latest main and not be ready for its contents.

For example, I recently created a merge request in the app to resume
downloading readings, which required changes to the rustfk library, so
when I pushed those changes to the library the pipeline for the main
branch started failing because it was building against the latest version.
After this change, we'll have to change the sha revision of the dependency
and include that change in the merge request with other changes.

It does make local development a little clumsier because you'll have to
edit the Cargo.toml to point to your local version of the library and
be careful about checking that in.