Fixing iOS builds for latest XCode.
I've tried to document the issues I ran into here and any commands
that may be necessary to run on build servers or developer machines.

The main error I was fixing was related to `DT_TOOLCHAIN_DIR` which
seems to be a Podfile fix, but I ran into other, older issues during
the process like the `libffi` thing which I've seen before.