Let flutter control more welcome screen sizing.
With the fixed sizes, on some iOS versions the screen would overflow off
the bottom. This fixed the issue for me on the physical iOS device
I tested with. If we need to fix other devices in the future we
it's worth understanding how we should be testing things like this
across multiple devices. I'm hoping in many other situations we can
get away with scrolling, but can't here.

Checked on: iPhone 15, Linux