Refresh tokens on authentication error.
Prior to this we'd only refresh tokens on application startup. This change
simplifies the error returned from the Rust library so we can easily tell
that an authentication issue is occuring and refresh the tokens. It also
applies that new error to any call that communicates with the portal.

Definitely could improve some of this rust code, now that I'm more
experienced with the language, haha