A single common ElevatedTextButton.
Our button styles were kind of all over the place and to try and wrangle things
in this enforces all buttons of this "kind" use the same widget. Then, we can
ensure consistency in style using the button style theming in `main.dart` and by
providing more useful contextual clues to the new widget.