Upgrading flutter_rust_bridge.
Lots happening here, as this includes the jump to the 2.0 branch.

Notable changes happening:

Common directories and files renamed, especially the generated code as well
as the common/standard paths to where the rust code lives in the project.
This means things moved from ./native to ./rust, introduced
a mysterious and poorly documented 'rust_builder' directory, and
meant many of the generated dart bindings were also renamed.

Another significant change is the generated dart code no longer
lived in a "Native" class, which meant the api object went away. This
created a new collisions between duplicate function names, so those had to
be disambiguated.

DateTime<Utc> stopped working in the bindings, so that became a new
custom type passing millis since epoch around.