Update xcode project to fix signing.
Ok, here's what I think happens.

The flutter build appears to build with the default settings in
the xcodeproject and then applies the export options plist. So if the
build won't work from xcode, using the settings there it'll fail before
ever getting to the provisioning profile you specify in ExportOptions

It may be possible to do one build with --no-codesign and *then*
another build with the export-options?

At any rate, the local changes matter so let's at least make them default
to something that works in the build. Developers can modify them locally.

Also, the environment varibales just confuse the pod dependencies
and opens up a whole other can of worms.

I take it as a testament to how screwed up Apple's sytem is that all the
OSS projects I've ever dealt with stumble through this like a drunk idiot
in the dark.

Oh, and this build definitely has a 50/50 chance of failing.